Authentication now works

+++ crypto/include/sha256.h
@@ -14,5 +14,5 @@
 #include <stdint.h>
 #include <crypto_utils.h>
 
-CRYPTO_ERR hash_sha256(uint8_t *input, size_t input_len, uint8_t *salt, uint8_t salt_len, uint8_t *digest);
-CRYPTO_ERR verify_sha256(uint8_t *input, size_t input_len, uint8_t *salt, uint8_t salt_len, int32_t fd_dgst);
+CRYPTO_ERR hash_sha256(uint8_t *input, size_t input_len, uint8_t *salt, uint32_t salt_len, uint8_t *digest);
+CRYPTO_ERR verify_sha256(uint8_t *input, size_t input_len, uint8_t *salt, uint32_t salt_len, int32_t fd_dgst);
